move data outside pages
